pyproject.toml
[project]
name = "genaiproject"
version = "0.1.0"
description = "AI Product Listing Generator: image-to-text, structured metadata, and asset pipelines"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.2",
    "httpx>=0.27.0",
    "Pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "typer>=0.12.5",
    "rich>=13.8.1",
]

[project.scripts]
genai-listing = "genaiproject.cli:app"

[tool.uvicorn]
app = "genaiproject.api.main:app"
